Autorizador_ABC/requirements.txt
pycryptodome==3.20.0
mysql-connector-python==9.0.0
